0.1 only prot of _putchar in hdr

+++ 0x14-more_singly_linked_lists/0-print_listint.c
@@ -1,4 +1,5 @@
 #include "lists.h"
+#include "stdio.h"
 /**
  * print_listint - print the str of linked list and return values in list
  * @node: starting node to print in list
@@ -6,7 +7,7 @@
  */
 size_t print_listint(const listint_t *node)
 {
-	int i = 0;
+	size_t i = 0;
 
 	while (node)
 	{
